requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Flask==2.2.2
Flask-WTF==1.0.1
future==0.18.2
geographiclib==1.52
geopy==2.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-decouple==3.6
ratelim==0.1.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
